move collections v2 code into versionless directories
