+++ client/src/plus/collections/index.scss
@@ -1,4 +1,4 @@
-@use "../../../ui/vars" as *;
+@use "../../ui/vars" as *;
 
 .collections-v2 {
   padding: 1rem 0;
